Refactor tests into separate folder
