+++ jest.config.js
@@ -152,9 +152,7 @@ module.exports = {
   // ],
 
   // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
-  // testPathIgnorePatterns: [
-  //   "\\\\node_modules\\\\"
-  // ],
+  testPathIgnorePatterns: ["__tests__/helpers"],
 
   // The regexp pattern or array of patterns that Jest uses to detect test files
   // testRegex: [],
